Extract utils and delete file: `src/utils/parsers/index.js`
