+++ src/app/core/layers/mixins/geo.js
@@ -5,11 +5,11 @@
  * @since 3.9.0
  */
 
-import GUI from 'services/gui';
-import ApplicationService from 'services/application';
+import GUI                               from 'services/gui';
+import ApplicationService                from 'services/application';
+import { getScaleFromResolution }        from 'utils/getScaleFromResolution';
 
 const Projections                        = require('g3w-ol/projection/projections');
-const { getScaleFromResolution }         = require('utils/ol');
 const { createFeatureFromFeatureObject } = require('utils/geo');
 const { XHR, sanitizeUrl }               = require('utils');
