+++ src/app/core/layers/providersfactory.js
@@ -2,6 +2,8 @@ import ApplicationState            from 'store/application-state';
 import RelationsService            from 'services/relations';
 import { QUERY_POINT_TOLERANCE }   from 'constant';
 import { QgsFilterToken }          from 'core/layers/utils/QgsFilterToken';
+import { getDPI }                  from 'utils/getDPI';
+import { getExtentForViewAndSize } from 'utils/getExtentForViewAndSize';
 
 const G3WObject                    = require('core/g3wobject');
 const {
@@ -17,12 +19,11 @@ const {
 const Parsers                      = require('utils/parsers');
 const { t }                        = require('core/i18n/i18n.service');
 const Feature                      = require('core/layers/features/feature');
-const geoutils                     = require('utils/ol');
 const Filter                       = require('core/layers/filter/filter');
 
 
 const GETFEATUREINFO_IMAGE_SIZE = [101, 101];
-const DPI = geoutils.getDPI();
+const DPI = getDPI();
 
 const is_defined = d => undefined !== d;
 
@@ -507,7 +508,7 @@ const Providers = {
         ? layers.map(layer => layer.getWMSInfoLayerName()).join(',')
         : this._layer.getWMSInfoLayerName();
 
-      const extent = geoutils.getExtentForViewAndSize(coordinates, resolution, 0, size);
+      const extent = getExtentForViewAndSize(coordinates, resolution, 0, size);
 
       const is_map_tolerance = ('map' === query_point_tolerance.unit);
