+++ src/app/g3w-ol/controls/basequerypolygoncontrol.js
@@ -3,9 +3,9 @@
  * @since v3.8
  */
 import { SPATIAL_METHODS } from 'app/constant';
+import { mergeOptions } from 'utils/mergeOptions';
 
 const InteractionControl = require('g3w-ol/controls/interactioncontrol');
-const { merge }          = require('utils/ol');
 const { Geometry }       = require('utils/geo');
 
 const VALIDGEOMETRIES    = Geometry.getAllPolygonGeometryTypes();
@@ -32,7 +32,7 @@ const BaseQueryPolygonControl = function(options = {}) {
     onhover: true
   };
 
-  options = merge(options, default_options);
+  options = mergeOptions(options, default_options);
 
   options.geometryTypes = VALIDGEOMETRIES;
