+++ src/app/g3w-ol/controls/lengthcontrol.js
@@ -1,4 +1,5 @@
-const utils = require('utils/ol');
+import { mergeOptions } from 'utils/mergeOptions';
+
 const LenghtIteraction = require('g3w-ol/interactions/lengthinteraction');
 const MeasureControl = require('g3w-ol/controls/measurecontrol');
 
@@ -10,7 +11,7 @@ const LengthControl = function(options={}) {
     interactionClass: LenghtIteraction
   };
 
-  options = utils.merge(options,_options);
+  options = mergeOptions(options,_options);
   MeasureControl.call(this, options);
 };
