+++ src/app/g3w-ol/layers/rasters.js
@@ -1,4 +1,6 @@
-const DPI = require('utils/ol').getDPI();
+import { getDPI } from 'utils/getDPI';
+
+const DPI = getDPI();
 
 const RasterLayers = {};
