+++ src/utils/getDPI.js
@@ -1,0 +1,8 @@
+import { DOTS_PER_INCH } from 'constant';
+
+/**
+ * @returns { number }
+ */
+export function getDPI() {
+  return DOTS_PER_INCH;
+};
