+++ src/utils/mergeOptions.js
@@ -1,0 +1,3 @@
+export function mergeOptions(obj1, obj2) {
+  return { ...obj1, ...obj2 };
+};
