+++ src/utils/needUseSphereMethods.js
@@ -1,0 +1,9 @@
+/**
+ * @param projection
+ * 
+ * @returns { boolean } 
+ */
+export function needUseSphereMethods(projection) {
+  return 'EPSG:3857' === projection.getCode() || 'degrees' === projection.getUnits();
+  //return projection.getUnits() === 'degrees';
+};
