+++ src/utils/removeMeasureTooltip.js
@@ -1,0 +1,16 @@
+/**
+ * Remove mesure tootltip
+ * 
+ * @param { Object } opts
+ * @param opts.map
+ * @param opts.tooltip
+ * @param opts.unByKey 
+ */
+export function removeMeasureTooltip({
+  map,
+  tooltip,
+  unbyKey,
+}) {
+  map.removeOverlay(tooltip);
+  ol.Observable.unByKey(unbyKey);
+};
