+++ src/utils/setMeasureTooltipStatic.js
@@ -1,0 +1,8 @@
+/**
+ * @param tooltip 
+ */
+export function setMeasureTooltipStatic(tooltip) {
+  const element = tooltip.getElement();
+  element.className = 'mtooltip mtooltip-static';
+  tooltip.setOffset([0, -7]);
+};
